+++ src/HabiticaAPIManager.js
@@ -177,6 +177,40 @@ class HabiticaAPIManager {
 		return promise;
 	}
 
+  /**
+	 * Make a POST request to the Habitica API.
+	 * Data object returned varies based on the API url called.
+	 * @param {string} baseURL - the url of the api call.
+	 * @param {HabiticaUserID} userID - the ID of the user, needed for authentication.
+	 * @param {HabiticaAPIToken} userAPIToken - the API Token for the user, needed for authentication.
+	 * @param {object} [queryParams={}] - key-value pairs for any parameters needed by the api call.
+	 * @returns {Promise<String>} Promise containing the raw API response data as a string.
+	 */
+	postRequest(baseURL, userID, userAPIToken, queryParams={}) {
+		let promise = new Promise((resolve, reject) => {
+			let req = new XMLHttpRequest();
+			req.open("POST", baseURL);
+
+			req.onerror = function() {
+				reject(this.responseText);
+			};
+
+			req.onload = function() {
+				if (req.status === 201 || req.status === 200) {
+					resolve(this.responseText);
+				} else {
+					reject(this.responseText);
+				}
+			}
+      req.setRequestHeader("Content-Type", "application/json;charset=UTF-8");
+			req.setRequestHeader("x-client", this.xclient);
+			req.setRequestHeader("x-api-user", userID);
+			req.setRequestHeader("x-api-key", userAPIToken);
+			req.send(JSON.stringify(queryParams));
+		});
+		return promise
+	}
+
 	// CLASS HELPER FUNCTIONS
 
 	/**